backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.0
tenacity==9.0.0
openai==1.58.1
celery[redis]==5.4.0
redis==5.2.0
azure-cosmos==4.9.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
